Remove option dired-du-duc-directories
It turns out that "duc ls" is not smart enough to look up the
DB for a sub-directory of an indexed directory anyway.

Which means we'd have to recurse into hundreds of thousands of
directories and explicitly run "duc index" on each --
so no.

Instead, the recurring timer will just re-index those directories that
have been visited by user at any point in current session.